qa/admin: next round of upgrade-test scripts
This is now complete and working for (at least) RHEL 8,
and previous version worked for .deb distros (but will need to
be revisited after the rpm changes).

Use (the new) upgrade-setup to get going (rpm only at the moment),
this will call get-distro-packages if needed).

Then upgrade-test will do the job, as in ...
 - if <distro> packages are available locally below $home/build
   + remove all PCP packages
   + install the <distro> packages
   + upgrade to the <latest> developer packages
   + run qa/check -g sanity
 - if <previous> developer packages are available locally below $home/build
   + remove all PCP packages
   + install the <previous> developer packages
   + upgrade to the <latest> developer packages
   + run qa/check -g sanity