libpcp: change in archive value interpolation for non-counter metrics
There has been some discussion here
    https://github.com/performancecopilot/pcp/issues/433
and here
    https://github.com/performancecopilot/pcp/pull/2078
about interpolation mode and non-interpolation mode for PCP archives.

I remain convinced that interpolation mode is preferred in almost
all circumstances, however I've been persuaded (thanks Marko for
resurrecting an issue that I've been neglected for almost 7 years) that
a change to the algorithms is warranted in the case of instantaneous
(and discrete) metrics ...

Currently we *extrapolate* from the prior observation (last seen).

The proposal is to use the *nearest* (in time) observation.

This change will be release in PCP 6.3.2 (due for release in about
3 weeks).

A full explanation of the new (and the old) algorithm can be found
in the soon to be updated version of
    https://pcp.io/papers/archive-interpolation.pdf