pmlogcheck: speed up, part 1
Thanks to Sagar (sagar.sagar@oracle.com), some slowness in pmlogcheck
with large archives has been observed.

See also: https://github.com/performancecopilot/pcp/issues/2082

This commit removes the O(N*N) part of metric-instance lookup in
pass3(), and replaces it with O(N) a [nested] hash table lookup.

There may well be a part 2, where we'll be able to decompress the
archives just once (at the cost of extra disk space while pmlogcheck
is running, which is why it will be a non-default option).  This commit
includes src/pmlogcheck/logdecompress (new) to decompress the physical
files of an archive.