Assorted: add support for zstd(1) compression of archives
These changes enable zstd(1) to be used via $PCP_COMPRESS
in pmlogger/pmie control files, or in the environment for the
pmlogger_daily and pmie_daily scripts, or as the -X command line
argument to the pmlogger_daily and pmie_daily scripts.

There is no change (yet) to the default use of xz(1) as the compression
tool.

Source code changes:
   src/libpcp/src/io.c
   src/libpcp_qed/src/qed_fileiconprovider.cpp
   src/pmafm/mkaf
   src/pmie/pmie_daily.sh
   src/pmlogcheck/logdecompress
   src/pmlogger/pmlogger_daily.sh
   src/pmlogger/pmlogger_daily_report.sh
   src/pmlogger/pmlogger_janitor.sh
   src/pmlogger/pmlogger_rewrite.sh
   src/pmlogger/utilproc.sh

Man page changes:;
   man/man1/pmie_check.1
   man/man1/pmlogger_daily.1

QA changes:
   some modified tests and a bunch of new ones