qa/1570: zstd DOES need --no-progress, sometimes
So on CentOS Stream9 (vm26 and in CI), zstd -q does not quieten the
progress meter babble.

This commit fixes qa/1570, but I'll have to circle back round and
rework the uses of zstd in libpcp, utilproc.sh (for pmlogger scripts)
and pmlogcompress.

Unfortunately this other work won't make the next release.