pmlogmv: rework and new pmlogcp command
There are lots of cosmetic changes in pmlogmv.c associated with
renaming the command line arguments: oldname, newname ->
srcname, dstname.

But the essential changes are:

1. Allow dstname to be an existing directory and use the basename of
   srcname as the basename of the destination archive

2. Use argv[0] to switch between pmlogmv (move, rename) and pmlogcp
   (copy).

So pmlogmv and pmlogcp now look like and behave much like mv(1)
and cp(1).

A whole bunch of changes in QA to use pmlogcp (in particular) instead
of cp(1), and thereby no longer relying on the shell's glob expansion
to enumerate the files of an archive.

[note to self] one less itch to scratch!