pmlogger maintenance scripts: switch to pmlogcompress
This means for most users we'll change under the covers from defaulting
to xz(1) to defaulting to zstd(1).

The trigger for doing immediate (and unconditional) compression
has been softened ... if zstd(1) is installed, we're doing it else
(old rule) if xz(1) is installed _and_ liblzma is available, we're
doing it ...  so this might mean compression for some where it was
not happening by default before.

This is CHANGELOG-worthy, but it is not expected to be visible to
users unless they're explicitly looking for .xz extensions.