qa/common.secure, qa/713: checks for readable path to qa/tls.conf/...
This test requires pmcd running as user $PCP_USER to be able to read QA
files in the qa/tls.conf directory, so add checks here and warn loudly.

Note that if this is not the case, qa/713 sued to fail in a miserable
way that provided no clue as to the root cause of the failure.

Most common root cause of failure here is running QA from the account
of some user for whom $HOME has been set up in the more modern
paranoid style as mode 700 rather than the older more permissive
style as mode 755.

Probably need to refactor this test to copy the required files into
someplace $PCP_USER is sure to be able to read.