configure.ac: fix unreliable tests on scandir() and struct dirent
These two tests should not rely on whether the compiler emitted
warnings, but should instead rely on whether the compilation failed.

For the scandir() test, with some combination of CFLAGS, the compiler
may emit some warnings, failing the test, but in reality the scandir()
call must use `const struct dirent *' for the callbacks.