pmlogger_merge: optionally compress output archive
Once the merge has been successfully completed, if any of the input
archive files was compressed, then compress the output archive.

With the change to zst, the highwater mark for disk space is
potentially much higher when pmlogger_daily tries to catch up multiple
days of housekeeping ... this change reduces (but does not remove)
the chance of anything bad happening in the process (like filling
the /var filesystem, as I've seen on bozo and vm01).