qa/1705: rework number of traces limit test (test4)
At some point, the semantics of "number of probes" changed, and
kprobe:* that expands to a bazillion of tracepoints that was more than
enough to exceed the default 512 limit is now counted as ONE probe.

Enumerate 513 explicit tracepoints to force the test to fail (as
expected) with both the old and new semantics.