bind2 PMDA: fix PMID assignments
Earlier commit had a few botches in the assignment of the PMID's
item field to known metrics.

Fix these, and harden mk.rewrite, so that in addition to generating
the pmlogrewrite config, this script now also
- reports error and exit status 1 for
  + duplicate metric names
  + duplicate PMIDs
- reports warning and exit status 0 for
  + gaps in the range of assigned item field values

If these had been in place, the earlier (wrong) changes in pmdabind2.pl
would have been caught with build failures.