libpcp: replace strncpy() and strncat() uses
I think it was in 1980 that I first learned about these routines in
Edition 7 (or possibly DWB) Unix.

It has taken 45 years, but I finally realize just how useless and
evil they really are, thanks in part to stumbling across this
    https://devblogs.microsoft.com/oldnewthing/20050107-00/?p=36773
during a completely unrelated fishing expedition with Google.

For reasons of history, not sanity, there is no good alternative that
is universally available in libc, so we had already created "safe"
versions in libpcp (pmstrncpy() and pmstrncat()) that ensure the
destination buffer is null-byte terminated, have a consistent use of
the "length" parameter and provide an indication of truncation via the
return value.  But these were not being used in many places in libpcp.

String copying is more common than string concatenation, and for one
of the strncpy() abuse cases, when the length of the source is known,
like in length-prefixed values in PDUs and on-disk data structures
for PCP archives, memcpy() + null-byte termination is a safe and
faster alternative, otherwise pmstrncpy() is the preferred solution.

This commit "fixes" libpcp by removing all uses of strncpy() and
strncat(), makes the man pages match the code and some related QA
issues.  Later commits will make the same style of changes across
other parts of the PCP code base.