mmv PMDA: rework diags
The mmv PMDA has reasonably robust checks when parsing the memory
mapped files that are producing the metrics, metadata and values,
but when something is NQR in the checking the PMDA tends to silently
ignore some or all of the memory mapped file.

In this commit, all hard check failures generate Error lines in the
log (it is pmcd's log, because the mmv PMDA is default configured
as a DSO agent) and soft check failures generate Warning or Info
lines in the log if -Dappl0 is in effect.

Additional changes include
- adding mmv.control.metrics to export the number of metrics
  successfully found in the memory mapped files
- small tweaks for mmvdump
- fix QA fallout after logging message changes
- new qa 1552 to bring back the long-forgotten qa/src/mmv_ondisk
  test application