CI: fix runner ubuntu-latest for F40 and Cent0S-Steam9
After wading through the podman-run(1) man page and only finding a
reference for --security-opt=no-new-privileges for *disabling* the
ability of a container to gain additional privileges, I started in
the Google swap where, on this topic in particular, the ignorance
vastly outweighs the wisdom.

Eventually (like hours later!) I stumbled on a "works for me" fix
that mentioned: --security-opt no-new-privileges=true, aha the old
double negative trick, so maybe we need
    --security-opt no-new-privileges=false

First good sign is the podman-run does not choke on this.

But problem (no user, not even root(!), can run sudo within the
container) persists.

After more hours of trial and error I accidently found that using both
--privileged and --security-opt silently does not work.  Dropping the
--privileged and BINGO --security-opt no-new-privileges=false takes
effect.

This commit makes that change in build/ci/ci-run.py, but only when
the runner is Ubuntu (the only case I can really test, but the one
we're interested in for CI).

And puts back os:ubuntu-latest for all platforms in CI.