libs: replace strncpy() and strncat() uses [part 2]
Replaced calls in these files:
    - libpcp_import/src/import.c
    - libpcp_mmv/src/mmv_stats.c
    - libpcp_pmda/src/pduroot.c
    - libpcp_pmda/src/root.c
    - libpcp_qed/src/qed_timecontrol.cpp
    - libpcp_trace/src/ftrace.c
    - libpcp_trace/src/p_data.c
    - libpcp_web/src/discover.c
    - libpcp_web/src/http_client.c
    - libpcp_web/src/json_helpers.c

Added libpcp_qed/install-dev and libpcp_web/install-dev as developer
"helper" scripts to do an "overlay" install over an existing PCP
install and build tree ... these libraries are not DSOs, so we need to
enumerate all the binaries (src and qa) that are linked with the .a
archives and rebuild (make <exec>) and reinstall (sudo make install)
'em.  This is useful as a quick short-cut to being able test changes
without the delays of a full Makepkgs-and-install cycle.

The install-dev scripts are modelled (in semantics, not form) on the
existing src/python/install-dev script.