qa/admin/pcp-daily: small interop testing change
vm16 runs PCP 4.0 for version interoperability testing, and pmlc
from PCP 4.0 to a current pmlogger won't work due to a change in the
protocol, unless the current pmlogger is running with the -I 1 option.

So skip qa/381 on vm16, 'cause this is going to fail when it picks
a remote QA host that is likely to be running a current version
of pmlogger.