pmwebapi: Implement filering for pmwebapi metric request
The pmwebapi metrics request can specify specific metrics from the
performance metrics name space (PMNS) to include in the response with
the names parameter.  In some cases is useful to filter or prune names
from the response.  The new filter parameter accepts a comma-separated
list of metric names to exclude in the response.  The names in the
list can include wildcards '*' and '?' for more generalized matches.