libpcp: add PM_CTXFLAG_LAST_VOLUME flag for pmNewContext()
When the context is an archive, this flag positions the archive at
the start of the last volume, instead of the start of the first volume
(the default).

Most useful for pmquery in "discovery" mode via libpcp_web, where
we're after the last timestamp in the archive and if the archive
is being concurrently written, the earlier volumes may have been
compressed so this change allows us to dodge decompressing volume 0.

Changes for pmapi.h and libpcp are included, along with using the
new flag in src/libpcp_web/src/discover.c.

The commit also provides man page updates and qa/1553 (new) and
qa/src/newcontext.c (new).