debian build: force debian/rules to be recreated each time
A build failure at Debian for i386 was triggered by configure correctly
picking that Perl is not Y2038 safe, but the associated rules.pcp
-> rules changes were not made because the GNUmakefile included a
dependency that was not deterministic (and failed in this case).

Replace debian/rules by a stub that will trigger early failure if
it is ever used (IIRC we need this to pass some integrity check
on the uploaded source), and force the GNUmakefile to unconditionally
recreate debian/rules from debian/rules.pcp each time.

This needs to be in a PCP 6.3.4 for Debian.