configure, images: always package desktop icons in pcp-gui
As reported at https://appstream.debian.org/sid/main/issues/pcp-gui.html
we are not (always) packaging the desktop icons for pcp-gui.

The icons are all checked into git, so they should be packaged.

The problem is the configure setting for HAVE_HICOLOR_ICONS which
was dependent on inkscape(1) being installed and inkscape is not
(and should not be) in the Build-Requires: list ... inkscape is
only needed to remake the icons and this is not required in a regular
build.