perl PMDAs: [lintian] be consistent about mode and (no) #! line
Although these are Perl scripts, they are never executed directly
from a shell, so mode 644 is more sensible than 755 (in the tree,
they are all installed 644) and there should NOT be a
	#! /usr/bin/perl
(or similar) line at the beginning of the file.

Making this consistent makes this sort of lintian warning:
script-not-executable [usr/lib/pcp/pmdas/bind2/pmdabind2.pl]
go away.