man pages: [lintian] be consistent about countrol for embedded tbl
man pages that include .TS ... .TE need pre-processing by tbl(1).

Some versions of man(1) know about this, others need "hints".

We had
- no hints in some files
- hints of the form
  '\"! tbl | mmdoc
  or
  '\"macro stdmacro
  that probably stopped being useful about the time SGI Irix went away
- hints of the form
  '\" t
  (as documented in the Linux man(1) page, but hidden in DEFAULTS
  and the description of "valid preprocessor string" that cycles
  back to the -p command line option)

Only the latter seems to shut lintian up, and the Linux version of
man(1) and the formatting of man pages in the build seems to get this
right irrespective of the "hints", so be consistent and use the form
    '\" t
that keeps lintian quiet.