qa: annotate test qa/1264 as 'not_for_ci'
This test started failing with the recent CI container platform upgrades,
and passes everywhere when run locally.  Something to do with puython PID
changing (namespaces?) from eBPF POV ... I'm sure this is not a PCP issue
but a true root cause and/or workaround for CI has proved elusive (and is
not worth chasing further IMO).