qa/1290: valgrind hack for vm11 just got hackier
vm11 is a Debian 10.13 i686 platform (the i686 part is probably
important), where valgrind is a bit odd.

We already had valgrind exceptions there for vm11, but when libpcp_web
has been stripped, the symbols are missing for the backtrace ... make
the valgrand exception backtraces even weaker to accommodate this.