archive-analysis container: calculate archive time range
This extends the PCP archive detection logic in the archive
analysis container such that we're now tracking the minimum
start and maximum end times of all archives detected.  This
may be possible to inject into our canned Grafana dashboard
(dashboards/pcp-archive-analysis.json) in the time.from and
time.to JSON fields.

Related to issue #2163