python api: dynamically adjust the timespec struct padding
Switch to adjusting the struct timespec ctypes padding at run
time instead of at build time.  This simplifies the build and
keeps the pmapi.py python code consistent.  Also solves a bit
of an LDIRT issue in the makefile with debian builds.  And we
remove the risk of accidentally losing WIP code changes when
someone hacking on pmapi.py doesn't realise it's generated.