CI: retire CentOS 6 and CentOS 7 platforms
Centos 6 EOL was 30 Nov 2020.
Centos 7 EOL was 30 Jun 2024.

Both are long gone, and there is no point wasting CI resources
building and testing the latest PCP on these platforms.

But the real trigger is that on github, the CentOS 7 container
depends on a version of Ubuntu that is EOL and we could not find
a way to make the CentOS 7 container there run correctly on a later
version of Ubuntu.