scripts/man-lint: add checks for test following .SH NAME
The gmake $(INSTALL_MAN) macro from builddefs has some (rather
strange) assumptions about the formatting of the text in this
section and in particular where the file name for the man
page is to be found ... if it cannot be found the man page
is SILENTLY not installed (!).

This change defends against this, and used pmdarocestat.1 as the
guinea pig test case.