build: [debian] clean up missing Build-Depends: packages
If we conditionally build a compenent (typically a PMDA) and the
guard checks for some file or module called "foo", then we need to
find the Debian package that installs "foo" and ensure that package
is named in Build-Depends: in debian/control.pcp.

Otherwise (as in the current PCP 6.3.8 in Debian trixie) we have a
bunch of PMDAs that are not built and packaged because "foo" is not
present on the build system.

This commit fixes a number of long-standing Debian bugs related to
this issue, and (more importantly) delivers a (new) qa/1580 that checks
for consistency betweeen the conditional use of $PMDA_FOO in the PMDA
GNUmakefiles, the presence of $PMDA_FOO in debian/fixcontrol and the
?{foo} pseudo-macros used in debian/control.* and debian/fixcontrol.