pmlogger, pmproxy: archive "push" functionality
Extension to the pmlogger utility to allow remote PCP archive
creation using a push mechanism, based on HTTP and a PMWEBAPI
/logger API, instead of the traditional pmcd pull mechanism.

Archives can be pushed to a remote pmproxy server as they are
being created by pmlogger, instead of (or as well as) written
to local disk, using the -R/--remote option.

A pmlogpush(1) utility is added to perform this 'remote push'
functionality for existing archives.

This commit includes libpcp_web improvements to its existing
HTTP GET operation (incl. transition of existing PMDAs using
this interface) and addition of a consistent POST operation,
which is used by pmlogger and pmlogpush.

Resolves #2059