libpcp & QA: address a very old bug
In the struct timespec conversion for pmSetMode (part of the
PMAPI_VERSION_4 work) I stumbled across a qa test app that was not
in the git tree, but had been developed to exercise this bug:
    https://bugzilla.redhat.com/show_bug.cgi?id=958745
[thanks Marko again]
but then abandoned.

So some TWELVE years later, the test was reactivited and lo and behold,
the original bug is no longer there, but it exposed a "been there
forever" bug in libpcp.  The bug involves changing archive read mode
mid-archive from PM_MODE_INTERP to either PM_MODE_FORW or PM_MODE_BACK
... furtunately no real PCP apps behave this way.

This commit delivers Marko's original reproducer archive (redacted
and compressed), the long-lost interp_bug3.c test app, some qa script
wrappers for non-valgrind and valgrind exercising and the one-line
libpcp fix.