PMAPI_VERSION_4: pmResult and pmFetch
This is the really big change in terms of scope and risk.

pmResult now contains a timespec, not a timeval, and pmResult_v2
provides the old definition.

pmResult_v2 is aliased (via #define) to pmdaResult and qmcResult
to provide ABI compatibility for PMDA callbacks and libpcp_qmc
clients.

Lots of consequential changes, e.g. pmFreeResult redefined to use
the new pmResult, pmFreeHighResResult (deprecated) and pmFreeResult_v2
added to support pmResult_v2.

Ditto for __pmOffsetResult and __pmOffsetHighResResult.

Ditto for __pmFreeResultValues and __pmFreeHighResResultValues.

Ditto for __pmDumpResult and __pmDumpHighResResult.

And uses of pmStore() rewritten to use pmStoreHighRes() ... this is
an interim step, as pmStoreHighRes() will be deprecated in the
next large commit in this series.