qa/src/test_pcp.python: et al
During the PMAPI_VERSION_4 work, qa/702 and qa/707 completely missed
the trashing of the pmResult due to a bogus offset in the Python
interface code.

Rework the Python part of the test to make the output more
human-friendly, and change 702 and 707 to emit filtered output to *.out
(not just the test's view of the end result).