PMAPI_VERSION_4: print and parse into {timeval,timespec} structs
For pmapi.h:
    pmPrintStamp -> pmtimevalPrint
    pmPrintHighResStamp -> pmtimespeecPrint

also
    pmPrintInterval -> pmtimespecPrintIntrerval
but this one was added after the last release, so this is a rename
before PCP 7.0 escapes.

For libpcp.h:
    __pmParseTime -> __pmtimevalParse
    __pmParseHighResTime -> __pmtimespecParse

These names are consistent with the names used for artithmetic
on the two structs, see man pmtimespecNow(3) for example, and
get rid of the ambiguus "Stamp" in the names of the routines.