PMAPI_VERSION_4: debug options fallout
So, changing the metrics that export debug options from type 32
to type STRING was such a good idea, ... except that if any of these
metrics are in archives, then was cannot merge a pre-PCP-7.0 archive
with a PCP-7.0 archive.

So, pmlogrewrite needs extending to allow <numeric> -> string type
conversion, and associated pmlogrewrite rules for the PMDAs that
are impacted.

But pmlogrewrite depends on pmExtractValue() and that needs extending
as well.

All of this needs some man page updates and additional QA coverage
with additional archives from vm32 where the archive merge snarfoo
was first observed.