PMAPI_VERSION_4: man pages for EventRecord services
This one is all a bit messy.

Because the PMAPI_VERSION_2 encoding of a pmResult (now known as
pmResult_v2) may exist in archives, or even older PMDAs, we can't
really make either the new or the old unpacking routines go
away, ditto for the freeing routines.

So we've resolved to keep all of:
    - pmUnpackHighResEventRecords() and pmUnpackEventRecords()
    - pmFreeHighResEventResult() and pmFreeEventResult()

But pmResult has changed and pmHighResResult is deprecated, so we
need to reword the man pages.