pmdaopenmetrics: use fine-grained timestamp for a stat-based guard
Switch from using stat(2) ctime in seconds to mtime in nanoseconds
on the openmetrics PMDA in an attempt to resolve a failure on some
of the QA farm machines.

This only affects the case where file-based metrics are sourced -
not the more usual cases of a REST API or script-based metrics.