qa/1976 et al: fix permissions issue leading to non-deterministic failures
The test was adding a URL that pointed to a file below $here ... if
the openmetrics PMDA running as the user $PCP_USER could not read
all the directories on the path to $here, the test failed, e.g. when
QA is run from ~mary/src/pcp/qa and ~mary has mode 700.

Move the target file and URL from $here to $tmp.

Also clean up the curl usage in the associated shell scripts so that
rc processing never happens and no progress meter output to mess up
$seq_full and make comparisons more difficult.