qa: ensure curl options used are ubiquitous
Fixes some fallout from commit de8ef436074 which introduced
the --no-progress-meter option in particular, which is only
available in newer versions of curl (and missing on some of
the older QA platforms currently supported).

Happily, -s achieves the same thing and is everywhere, so I
switched over to using that with the original option set.