libpcp_web: improve error handling in http return codes
Passing an invalid URL into pmlogger has been observed to
generate misleading error diagnostics.  This is because 1
or -1 was being used instead of meaningful code, and that
(1) maps to "Operation not permitted" irrespective of the
root cause.  Now we at least get EINVAL for incorrect URL
usage.  Some QA to follow with the new pmlogger push test
cases currently being added.