src/libpcp/src/context.c: audit error paths in pmDupContext() [Coverity]
The flow of control on the error paths in this routine is very
complicated, and Coverity has flagged a number of issues.  Fortunately
they mostly relate to malloc() failures, so the real risk is small.

To clean this up, fault-injection points have been added on all the
error code paths, and qa/1606 and qa/1607 exercise all of these
fault-injection points.

Hopefully this fixes CIDs 454672, 454660 and 454648