pmlogger_daily: fix $PCP_CULLAFTER non-implementation
An audit of pmlogger_daily prior to refactoring to extend support
for "push" archives revealed that although documented in the
pmloggger_daily(1) man page, $PCP_CULLAFTER (in the environment)
had never worked.

This commit replicates the logic for $PCP_COMPRESSAFTER (which does
work) to support $PCP_CULLAFTER.

The half-baked support for CULLAFTER (in a control file, but not in
the environment) is maintained by treating $CULLAFTER= as a synonym
for $PCP_CULLAFTER= in control files, but this is not documented
to encourage deprecation.