src/pmlogconf/pmlogconf.c: fix unsigned "underflow" [Coverity]
In update_groups() the loop control variable "i" is unsigned
('cause it is passed to another function), but this sequence
to "stay on the same group"
    for (i=0; ...) {
	...
	i--;
	continue
    }
is a bit naughty ... i-- underflows, then i++ at the end of
the loop overflows to get back to zero.

Make the logic a little clearer with another state variable.

Fixes CID 426724