Audit subsecond time reporting formats.
New scripts/check-time-formatting finds places where we report
subsecond times (usec or nsec) and tries to check if the associated
format uses %06d or %09d respectively.

The test is uses pattern-based heuristics (not language-specific
tools), so it is not foolproof.

Uses a meta-tag (comment) containing "check-time-formatting-ok"
ahead of a checkpoint to silence false matches.

Turns out the code is pretty good except for some false matches in
src/pmlogdump/pmlogdump.c and a couple qa apps.