WIP: libpcp_web: Address memory leak in webgroup_scrape_tree()
The strings in the scrape struct need to be cleaned out even in the
case that no matching metrics are found for them in the PMNS
(Performance Metric Name Space).

This fix doesn't appear to be quite right and might be papering over a
problem. There are definitely names in the scrape struct, but either
sts from pmTraversePMNS_r < 0 or one of the earlier runs of
webgroup_scrape_batch has set scrape->status < 0.