src/libpcp/src/logutil.c: rework for logFreeHashInDom()
pmproxy processing archives from a remote pmlogger exposed a
problem below pmDestroyContext() where V3 archive format uses
"delta" indom records, and the instances that don't change
between one indom observation and the next *share* a pointer
to the same malloc'd external instance name.

The free'ing logic was hitting double-free's as a consequence.

We don't have, and cannot add, reference counters for these pointers,
so the only option is to scan all of the external instance names for
each indom, removing (NULLing) the duplicate pointers.  Fortunately the
existing hashing routines can be leveraged to make this (potentially)
O(N*N) operation cheaper (and N might be big for, er, the indom for
process metrics).

We've not seen this in the past, because no other application has to
deal with incoming V3 delta indom records *and* call pmDestroyContext()
later on for the same PMAPI context.  This also means the cost of the
additional processing below pmDestroyContext() is expected to be in
the range "zero to undetectable".