src/pmlogger/src/callback.c: check_inst() fooled by derived metrics
We do not encourage logging derived metrics, but it is supported.

When doing some ad hoc testing of remote log pushing to pmproxy I
configured the pmlogger to log all Linux metrics, e.g. disk, and this
pulled in all the derived metrics with disk.* names.

Then for *every* pmResult, check_inst() was complaining that the
raw PMID was not found in the previous result (because it had been
mapped to a different PMID as per SET_DERIVED_LOGGED()) ... this led
to a huge pmlogger.log and surprises in the tail pmlogger.log window
where I was watching the life of pmlogger.

Make the check aware of the possible mapping change to the PMID and
the babble from check_inst() went away.