src/libpcp/src/logmeta.c: fix error status from __pmLogLoadMeta()
This has bugged me for a long time, and eventually was tripping up
pmlogger_daily so chronically in my test environment with lots of
archive churn.

The problem is most simply demonstrated by this anomaly with the
archive empty that contains only the archive label record:

    $ pmlogdump empty
    pmlogdump: Cannot open archive "empty": Empty archive file
    $ pmlogrewrite -i empty
    pmlogrewrite: Warning: empty archive "empty" will be skipped
    $ pmlogrewrite -qi empty
    pmlogrewrite: Error: cannot open archive "empty": Corrupted record in a PCP archive

And the trigger is that pmlogrewrite with -q opens the archive initially
with pmNewContext and the PM_CTXFLAG_METADATA_ONLY flag ... and this code
path returns PM_ERR_LOGREC not PM_ERR_NODATA when the metdata file is
empty.