pmlogger-pmlc: fix day-1 bug
Recent changes to delay preamble writing from pmlogger until the
first real pmResult is written led to a cascading series of events
that probably warrant a more detailed commit message to engage the
historians and archaeologists.

Because the output archive is now empty until the first pmResult is
written, several QA tests that used the
    pmlogger -L -c /dev/null ...
(or similar) pattern to get an archive started needed to be tweaked,
typically by using a pmlogger configuration file that contained
something like
    log mandatory on once sample.long.ten
which forces an early output of a small pmResult for one metric.

qa/061 then started displaying oddness and failure to return the
correct state in a pmlc query command.

Considerable effort was investigated in trying to identify the cause
of this apparent regression, but going back to PCP 4.0 (yep!) I could
not find anything that would explain the difference in behaviour.

The code in pmlogger that is supposed to manage the "hist_hash" of
logging state changes for all logged metrics and instances has no
debugging statements, so I added them (-Dappl8 for those who care).

A pmlogger optimization (from long ago) cut in and this was rolling the
two "log mandatory on once" metrics form qa/061 into one pmFetch and
one pmResult in the archive ... by luck (good or bad), sample.long.ten
was the first metric in the pmResult.

Eventually I discovered the metric that is the target of qa/061
(sample.colour) was *not* getting "hist_hash" updates, but the initial
sample.long.ten was getting the updates.

The root cause was a "return" that should have been a "continue"
in setavail() that meant the state on this code path was only ever
being set for the first metric in a pmResult, and the additional
metric pushed the metric qa/061 was looking for into second position
in the pmResult.

There are 50 QA tests that exercise pmlc, but qa/061 was the only
one failing.  When I fixed the code, the only additional failure
was qa/032 where the .out file contained one line that was in fact
wrong.

I've looked back to the earliest code I can find, and this bug has
been present since at least Sep 2002.