qa: cull some retired tests
The :retired tag in the groups file prevents these tests from being
run by default, but the contents of the tests confuses any attempt
for mass refactoring of tests as we're tricked into make changes in
tests that probably don't work anymore.

I don't want to git rm the files, as this would allow the test
sequence number to be re-used for something totally different, so
I've opted to replace the body of the test with
    _notrun "retired and culled"
so the original test remains in the git history, but the test
sequence number is also effectively retired.

This is the first batch of changes like this, triggered this time by
cleaning up the demise of "_service pcp" in QA tests.