qa: refactor for the demise of /etc/init.d/pcp
If you don't have systemd in the mix, then a big chunk of QA went
down the gurgler when /etc/init.d/pcp went away (recently).

This commit expunges support for, and all references to,

	_service pcp [stop|start|restart]

in QA land, and each of
	_service pmcd
and
	_service pmlogger
are [stop|start|restart]ed explicitly as needed.