Enable cron-driven work for Linux systems without systemd
Linux MX for example has systemd but it is not necessarily active.

We were packaging and installing all the systemd-driven timers but
none of the cron-driven configuration.  We can't safely install both
in the right place, and systemd(1) is going to win here.

So the crontab entries are installed in /usr/share/pcp/cron.d which
is safely hidden from cron(1), with a README that explains how to
install these crontab enties if needed.

This seems to have got the polling and daily processing working
(for the first time) on vm13.

There are no changes for non-Linux systems or Linux systems using
systemd(1).