pmproxy: drop the systemd unit dep between pmproxy and pmcd
This was recently added, but there are a number of pmproxy
deployment scenarios (incl. the archive-analysis container)
where no such dependency is required, nor wanted.

The downside of this is the user must be aware of any such
dependency, potentially.  In practice, deployments where a
pmlogger push pmproxy server is present, is very likely to
have pmcd running.  This is also the pcp-zeroconf default.