qa/119: re-arrange pmlogger restart logic
I don't understand this, but the sequence
	_service pmlogger restart 2>&1 | _filter_pcp_restart
	_wait_for_pmlogger
	_restore_auto_restart pmlogger
in _cleanup() was seeing the newly started primary pmlogger whacked
with a SIGTERM almost immediately.

Changing to this
	_restore_auto_restart pmlogger
	_service pmlogger restart 2>&1 | _filter_pcp_restart
	_wait_for_pmlogger
made the problem go away.

This has to be some instability between the test and systemd, but I
cannot see why.  The change in ordering may also be needed
(a) in the other tests that use _restore_auto_restart pmlogger, and
(b) and all the tests that use _restore_auto_restart pm<somethingelse>

But I'll wait until I see failures with a similar signature.