pmlogger + libpcp: fix lurking __pmOptFetchAdd bug
Bug has been there "forever", but Nathan's recent updates to pmlogconf
produced a new default pmlogger configuration file that added
	nfs.server.reqs
after the configuration file already had
	nfs.client.reqs[read,write]
and the former came *after* the latter in the configuration file.

When __pmOptFetchAdd() sees this it expands the instance profile for
the fetch group from 2 instances "read,write" to all instances, which
free'd an internal data structure that pmlogger had decided that it
was going to dink with, leading to a double free().  Smacking pmlogger
fixed the double free(), but the new qa/1631 that was the reproducer
exposed some more memory leaks lurking in __pmOptFetchAdd().