build: use rawhide (contains PCP v7) in the Containerfiles
Switch to using fedora rawhide instead of fedora:latest and
c10s in the two containerfiles.  This allows us to use the
most up-to-date version of PCP, which contains the components
we need like pcp-import-pmseries.

Using pre-built packages also solves a problem in CI where
the build (for multi-arch containers) takes >6hrs and ends up
cancelled as a result (github time limit reached).

This also helps to test latest PCP container images with the
latest changes arriving in rawhide.  One systemd fix relating
to the the recently removed sys-fs-fuse-connections.mount is
also included here as a result.

Another issue encountered relates to pcp-zeroconf in rawhide
generating a 'non fatal error' during rpm installation - and
microdnf/podman stops the build as a result.  For now I have
dropped pcp-zeroconf temporarily until that can be resolved.