containers: revert to a single source of truth for uid/gid
Debugging recent container changes I noticed a couple of problems
in the current containers due to recent changes in systemd.  One,
the sed workaround we had for pcp accounts is causing different,
conflicting values to exist for pcp uid/gid because it can only
be done after package install, which is too late.  Two, systemd
enters a wierd timeout state with its userdbd service as it tries
to reconcile a network view of users/groups with those inside the
container.  This caused slowness setting up PCP, due to the calls
to getpwent/getgrent by pmpost of all things.  I've switched this
additional systemd service off as its not needed here.