fixes for >2GB archive files on 32-bit platforms
We did the -D_FILE_OFFSET_BITS=64 thing a long time ago, but had
overlooked ftell() and fseek().

This commit changes all fseek() calls to fseeko() and ftell() calls
to ftello().

Also the pmLogTellCallBack typedef is changed to return an off_t
rather than a long.

There should be NO CHANGES on 64-bit platforms, and the only ones
to notice are likely to be 32-bit Debian or *BSD systems.