python api: improve exception handler in import API
While writing guidellm2pcp I encountered a double fault
(exception handler generating an exception) in the PMI
module, relating to failure to lookup a PCP API code in
the PMI error code list (which can happen for core PCP
errors, and pmiErrStr handles this just fine).

Failure case added to a new test 1735 to exercise this.